+++ server/cmd/information/system/casbin.go
@@ -92,6 +92,10 @@ var carbines = []gormadapter.CasbinRule{
 	{PType: "p", V0: "888", V1: "/workflowProcess/getMyStated", V2: "GET"},
 	{PType: "p", V0: "888", V1: "/workflowProcess/getMyNeed", V2: "GET"},
 	{PType: "p", V0: "888", V1: "/workflowProcess/getWorkflowMoveByID", V2: "GET"},
+	{PType: "p", V0: "888", V1: "/excel/importExcel", V2: "POST"},
+	{PType: "p", V0: "888", V1: "/excel/loadExcel", V2: "GET"},
+	{PType: "p", V0: "888", V1: "/excel/exportExcel", V2: "POST"},
+	{PType: "p", V0: "888", V1: "/excel/downloadTemplate", V2: "GET"},
 	{PType: "p", V0: "8881", V1: "/base/login", V2: "POST"},
 	{PType: "p", V0: "8881", V1: "/user/register", V2: "POST"},
 	{PType: "p", V0: "8881", V1: "/api/createApi", V2: "POST"},
