+++ server/model/exa_excel.go
@@ -1,0 +1,6 @@
+package model
+
+type ExcelInfo struct {
+	FileName string        `json:"fileName"`
+	InfoList []SysBaseMenu `json:"infoList"`
+}
