--- server/model/request/exa_file_upload_and_download.go
@@ -1,8 +1,0 @@
-package request
-
-import "gin-vue-admin/model"
-
-type ExcelInfo struct {
-	FileName string `json:"fileName"`
-	InfoList []model.SysBaseMenu `json:"infoList"`
-}
